+++ application/api/src/main/kotlin/io/raemian/api/goal/GoalService.kt
@@ -3,8 +3,6 @@ package io.raemian.api.goal
 import io.raemian.api.goal.controller.request.CreateGoalRequest
 import io.raemian.api.goal.controller.request.DeleteGoalRequest
 import io.raemian.api.goal.controller.response.CreateGoalResponse
-import io.raemian.api.goal.controller.response.GoalResponse
-import io.raemian.api.goal.controller.response.GoalsResponse
 import io.raemian.api.sticker.StickerService
 import io.raemian.api.support.RaemianLocalDate
 import io.raemian.api.tag.TagService
@@ -21,19 +19,6 @@ class GoalService(
     private val tagService: TagService,
     private val goalRepository: GoalRepository,
 ) {
-
-    @Transactional(readOnly = true)
-    fun findAllByUserId(userId: Long): GoalsResponse {
-        val goals = goalRepository.findAllByUserId(userId)
-        return GoalsResponse.from(goals)
-    }
-
-    @Transactional(readOnly = true)
-    fun getById(id: Long): GoalResponse {
-        val goal = goalRepository.getById(id)
-        return GoalResponse(goal)
-    }
-
     @Transactional
     fun create(userId: Long, createGoalRequest: CreateGoalRequest): CreateGoalResponse {
         val (title, yearOfDeadline, monthOfDeadLine, stickerId, tagId, description) = createGoalRequest
